tests/wpt/tests/websockets/constructor/010.html

<!doctype html>
<title>WebSockets: protocol in response but no requested protocol</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<meta name="variant" content="?wpt_flags=h2">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/handshake_protocol');
  ws.onopen = ws.onmessage = ws.onclose = t.step_func(e => assert_unreached(e.type));
  ws.onerror = t.step_func(function(e) {
    ws.onclose = t.step_func(function(e) {
      assert_false(e.wasClean, 'e.wasClean should be false');
      assert_equals(e.code, 1006, 'e.code should be 1006');
      t.done();
    });
  })
});
</script>
