Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/constructor/010-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/handshake_protocol' failed: Unexpected response code: 400

PASS WebSockets: protocol in response but no requested protocol 
